NS-1412 revise cache key strategy
This commit updates the cache key strategy in order to facilitate
clearing portions of the cache and decouple the cache clear strategy
from code in d360 that accesses the cache. Because api urls were
transformed into a MD5 hashs, the cache key task would be required
to rebuild all possible combiniations of path and query parameters
in order to invalidation a portion of the cache. The new solution
retains the url path and only hashes the parameters. Thus, all keys
matching a path can be simultaneously invalidated without knowing
the query parameters. Now the cache clear task will no longer need
to be updated as additional parameters are added to the api or
the usage of parameters changes in d360.