NS-1412: address explicit params and truncation
This commit updates the key generation strategy to handle the cases
where the :params option is supplied to ActiveResource, such as
it is used in nsight to access core-api V2 endpoints. It also
limits the path length to 150 characters so as to avoid exceeding
the FileStore file name length limits that could potentially occur
for very long endpoints.